# openapi.yml
openapi: 3.0.3

info:
  title: QueryRunner API
  version: "2.2.0"
  description: >
    QueryRunner API, SQL Server üzerinde güvenli ve kontrollü sorgu yürütmek için
    geliştirilmiş FastAPI tabanlı bir servistir.

    API, JWT tabanlı authentication, tarih aralıklı sorgular, statik sorgular,
    güvenli raw SQL (SELECT/WITH) çalıştırma ve MailLogger ile e-mail log
    gönderimi özelliklerini destekler.

    ⚠️ Güvenlik Notu:
    DDL ve DML komutları (INSERT, UPDATE, DELETE, DROP, ALTER, TRUNCATE vb.)
    otomatik olarak engellenir ve tüm denemeler MailLogger aracılığıyla loglanır.

servers:
  - url: http://192.168.2.98:8275
    description: Production Docker Server (On-Prem)

tags:
  - name: Auth
    description: JWT tabanlı kullanıcı doğrulama
  - name: Query
    description: Kayıtlı SQL sorgularını çalıştıran endpointler
  - name: Raw SQL
    description: >
      Güvenli SELECT / WITH tabanlı SQL yürütme.
      DDL/DML komutları otomatik olarak engellenir ve loglanır.
  - name: System
    description: Sistem ve sağlık kontrolü endpointleri

paths:
  /:
    get:
      summary: Health Check
      description: Sunucunun aktif olup olmadığını doğrular.
      tags: [System]
      responses:
        "200":
          description: API çalışıyor
          content:
            application/json:
              example:
                message: "QueryRunner API is running 🚀"

  /auth/login:
    post:
      summary: Kullanıcı girişi (JWT Token üretimi)
      description: >
        Başarılı giriş sonrası token **Bearer <token>** formatında döner.
        Bu token tüm korumalı endpointlerde Authorization header içinde kullanılmalıdır.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: test_user
                password:
                  type: string
                  example: test_password
      responses:
        "200":
          description: JWT token başarıyla üretildi
          content:
            application/json:
              example:
                access_token: "Bearer eyJhbGciOiJIUzI1NiIsInR..."
                token_type: bearer
        "401":
          description: Kullanıcı adı veya şifre hatalı

  /query/:
    post:
      summary: Tarihli veya tarihsiz kayıtlı SQL sorgularını çalıştır
      tags: [Query]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: string
                  example: ["Net_Satışlar"]
                Tarih:
                  type: string
                  description: "Örnek: 02.01.2025 ile 31.08.2025"
                  example: "02.01.2025 ile 31.08.2025"
      responses:
        "200":
          description: Başarılı
          content:
            application/json:
              example:
                user_id: 12
                items: ["Net_Satışlar"]
                result:
                  Net_Satışlar: 1500000
        "401":
          description: Token geçersiz

  /query/v2/run-basic:
    post:
      summary: Tarihsiz statik SQL sorgularını çalıştır
      tags: [Query]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key:
                  type: string
                  example: IK_Aylik_Ozet
      responses:
        "200":
          description: Sorgu başarıyla çalıştırıldı
          content:
            application/json:
              example:
                user_id: 12
                key: IK_Aylik_Ozet
                rowcount: 24
                data:
                  - Departman: "Sos İşletme"
                    PersonelSayisi: 18
        "404":
          description: SQL key bulunamadı
        "401":
          description: Token geçersiz

  /query/raw/run-sql:
    post:
      summary: Güvenli Raw SQL çalıştır (SELECT / WITH)
      description: >
        Bu endpoint yalnızca SELECT ve WITH (CTE) ifadeleri ile başlayan SQL komutlarını kabul eder.

        ❌ INSERT, UPDATE, DELETE, DROP, ALTER, TRUNCATE, EXEC, MERGE gibi
        DDL/DML komutları otomatik olarak engellenir ve MailLogger ile loglanır.
      tags: [Raw SQL]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sql]
              properties:
                sql:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  description: "Tek SQL veya SQL listesi. Sadece SELECT / WITH."
                note:
                  type: string
                  nullable: true
                  description: "Opsiyonel log açıklaması"
      responses:
        "200":
          description: Sorgu(lar) başarıyla çalıştırıldı
          content:
            application/json:
              example:
                user_id: 12
                results:
                  - index: 0
                    columns: ["Name", "Age"]
                    rowcount: 2
                    rows:
                      - { "Name": "Nilay", "Age": 28 }
                      - { "Name": "Ayşe", "Age": 29 }
        "400":
          description: SELECT/WITH/EXEC dışındaki SQL komutları yasaktır
        "401":
          description: Token geçersiz
        "500":
          description: Sunucu hatası

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT